Change url for keycloak and extracted the zipfile creation to a function
