+++ frontend/teacher-panel/src/routes/api/v1/teachers/[id]/assignment/logs/[assignment_id]/+server.ts
@@ -3,6 +3,7 @@ import{ eq, inArray } from 'drizzle-orm'
 import * as schema from '$lib/db/schema'
 import JSZip from 'jszip';
 import { login } from '$lib/api/login';
+import { zipSubmissions } from '$lib/api/utilities';
 
 
 /**
@@ -22,22 +23,8 @@ export const GET = async ({params, request}) => {
     }).from(schema.submission)
       .where(eq(schema.submission.assignment_id, assignment_id));
 
-    const zip = new JSZip();
-
-    for (let i = 0; i < submissions.length; i += 1) {
-        const submission = submissions[i];
-        const student_folder = zip.folder(`student_${submission.student_id}`);
-        if (!student_folder) {
-            return new Response(JSON.stringify({
-                message: `Failed to create folder for student ${submission.student_id}`
-            }), { status: 500 });
-        }
-        student_folder.file("stdout.txt", submission.stdout);
-        student_folder.file("stderr.txt", submission.stderr);
-    }
-
-    const zipFile = await zip.generateAsync({ type: "blob" });
-
+    const zipFile = await zipSubmissions(submissions);
+    
     return new Response(JSON.stringify({
         message: "Success",
         zipFile: zipFile
